회의실 예약 api 제작 - 해당날짜에 예약된 회의실 전체 조회 - 회의실 예약하기 - 예약한 회의실 목록 조회하기 - 회의실 예약 취소

+++ src/main/java/com/officehub/officehubserver/OfficeHubServer/dto/PostMeetingRoomReservationDto.java
@@ -10,8 +10,8 @@ import java.time.LocalDate;
 @Setter
 @Alias("postMeetingRoomReservation")
 public class PostMeetingRoomReservationDto {
-    private Long roomId;
-    private Long subscriberId;
+    private int roomId;
+    private int subscriberId;
     private int startHour;
     private int endHour;
     private LocalDate reservationDate;
